re-organize testing code, config jest and enzyme
