+++ __mocks__/testReduxIntWrapper.js
@@ -2,7 +2,7 @@ import {IntlProvider} from 'react-intl'
 import {MemoryRouter as Router} from 'react-router'
 import {Provider} from 'react-redux'
 import React from 'react'
-import translations from '../i18n'
+import translations from '../src/i18n'
 
 const ReduxIntWrappeComponent = (store, component, locale = 'fi') => {
     return (
