+++ src/views/Search/Search.test.js
@@ -1,47 +1,15 @@
 import configureStore from 'redux-mock-store'
 import React from 'react'
-import renderer from 'react-test-renderer'
+import {shallow} from 'enzyme'
 import thunk from 'redux-thunk'
 
-import initializeAppSettingsForJest from '../../utils/jestAppSettings'
-import testReduxIntWrapper from '../../utils/testReduxIntWrapper'
-initializeAppSettingsForJest()
+import testReduxIntWrapper from '../../../__mocks__/testReduxIntWrapper'
 import Search from './index'
+import {mockUser} from '__mocks__/mockData';
 
-// findDOMNode mock is needed because we use React 15 with Material-UI v1. When upgrading to React 16 this mock can be removed.
-jest.mock('react-dom', () => ({
-    findDOMNode: () => {},
-}))
-
-const mockStore = configureStore([thunk])
-let store
-
+const mockStore = configureStore([thunk]);
 const initialStore = {
-    user: {
-        id: '0a423e5a-d34a-11e7-9a41-c2a5d78378ac',
-        firstName: '',
-        lastName: '',
-        username: 'u-bjbd4wwtjii6pgsbyks5pa3yvq',
-        emails: [
-            {
-                value: 'foo@foo.com',
-            },
-        ],
-        provider: 'helsinki',
-        _raw: '{"last_login":"2018-01-19T07:33:42.102250Z","username":"u-ajbd4wwtjii6pgsbyks5pa3yvq","email":"foo@foo.com","date_joined":"2017-11-27T08:07:41.123918Z","first_name":"","last_name":"","uuid":"0a423e5a-d34a-11e7-9a41-c2a5d78378ac","department_name":null}',
-        _json: {
-            last_login: '2018-01-19T07:33:42.102250Z',
-            username: 'u-ajbd4wwtjii6pgsbyks5pa3yvq',
-            email: 'foo@foo.com',
-            date_joined: '2017-11-27T08:07:41.123918Z',
-            first_name: '',
-            last_name: '',
-            uuid: '0a423e5a-d34a-11e7-9a41-c2a5d78378ac',
-            department_name: null,
-        },
-        token: 'sometoken',
-        organization: 'ahjo:02100',
-    },
+    user: mockUser,
     events: {
         apiErrorMsg: null,
         isFetching: false,
@@ -50,12 +18,10 @@ const initialStore = {
         eventError: null,
         eventsError: null,
     },
-}
+};
 
 describe('Search Snapshot', () => {
-    beforeEach(() => {
-        store = mockStore(initialStore)
-    })
+    let store;
 
     it('should render view correctly', () => {
         const componentProps = {
@@ -65,10 +31,11 @@ describe('Search Snapshot', () => {
                 },
             },
         } // Props which are added to component
+        store = mockStore(initialStore);
         const componentToTest = <Search {...componentProps} />
-        const renderedValue = renderer.create(testReduxIntWrapper(store, componentToTest)).toJSON()
+        const wrapper = shallow(testReduxIntWrapper(store, componentToTest));
 
-        expect(renderedValue).toMatchSnapshot()
+        expect(wrapper).toMatchSnapshot()
 
     })
 })
